Duration and TimerPosition as double, create intial song if list empty

+++ RogueSound.Functions/SongCurrentModel.cs
@@ -8,6 +8,6 @@ namespace RogueSound.Functions
     {
         public string SongId { get; set; }
 
-        public int TimerPosition { get; set; }
+        public double TimerPosition { get; set; }
     }
 }
